# openapi.yaml
openapi: 3.1.0
info:
  title: Videa 外部 API
  version: 1.0.0
  description: |
    第三方 Web 应用使用 Videa 作为支付和分成引擎的外部 API。

    ## 身份认证

    两种认证模式：

    1. **Bearer + API Key** - 用于用户级操作（支付、余额、收益、事件）。传递 `Authorization: Bearer <token>` 和 `X-API-Key: videa_...` 请求头。
    2. **仅 API Key** - 用于平台级操作（收益方案）。仅传递 `X-API-Key: videa_...` 请求头。

    Bearer Token 通过 SIWE（Sign-In with Ethereum）OAuth 授权码流程获取。

    ## 核心概念

    - **代币**：USDT 和 USDC（6 位精度）。Decimal 字段序列化为字符串。
    - **链**：Polygon (137)、BSC (56)、Polygon Amoy (80002)、BSC Testnet (97)。
    - **收益方案**：BALANCED (70/20/10)、CHANNEL_PRIORITY (60/30/10)、STRONG_CHANNEL (50/40/10)、CREATOR_PRIORITY (80/10/10) — 创作者/分销商/平台百分比。
    - **幂等性**：支付路由（购买、订阅、打赏）需要 `Idempotency-Key` 请求头。
    - **时间戳**：ISO 8601 格式。

    ## 统一购买模型

    所有支付（内容购买、订阅和打赏）均存储为 **Purchase** 记录，
    通过 `category` 字段（`CONTENT`、`SUBSCRIPTION` 或 `TIP`）区分。`/subscribe` 和 `/tip`
    路由作为便捷别名存在，但会创建对应类别的 Purchase 记录。
    新集成建议使用 `POST /v1/external/purchase` 并指定所需的 `category`。

    ## 弹窗流程

    Videa 提供独立的弹窗页面处理支付操作，包括钱包连接
    （SIWE）、链/代币选择、余额检查和内联充值——无需父应用
    提供 Bearer Token。

    支付弹窗（购买、订阅、打赏）需要**支付意图**。通过
    `POST /v1/external/payment-intents` 在服务端创建，然后仅将 `intent_id` 传递给
    弹窗 URL。意图在 30 分钟后过期且仅可使用一次。

    | 弹窗 URL | 必填参数 | 成功事件 |
    |-----------|----------------|---------------|
    | `/purchase` | `client_id`, `redirect_uri`, `user_id`, `intent_id` | `videa:purchase:success` |
...
    | `/tip` | `client_id`, `redirect_uri`, `user_id`, `intent_id` | `videa:tip:success` |

...

...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 通过 POST /api/auth/token/exchange 获取的 Bearer Token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 用于应用识别的 API 密钥
  schemas:
    RevenueSharePlan:
      type: string
      enum: [BALANCED, CHANNEL_PRIORITY, STRONG_CHANNEL, CREATOR_PRIORITY]
...
    Token:
      type: string
      enum: [USDT, USDC]
...
    ChainId:
      type: integer
      enum: [137, 80002, 56, 97]
      description: "137=Polygon, 80002=Amoy, 56=BSC, 97=BSC Testnet"
...
    CreatePaymentIntentRequest:
      type: object
      required: [type, creatorUserId, revenuePlan, externalTransactionId]
      discriminator:
        propertyName: type
      properties:
        type:
          type: string
          enum: [purchase, subscribe, tip]
          description: 映射到购买类别——purchase=CONTENT, subscribe=SUBSCRIPTION, tip=TIP
        creatorUserId:
          type: string
        revenuePlan:
          $ref: "#/components/schemas/RevenueSharePlan"
        distributionLinkCode:
          type: string
        chainId:
          $ref: "#/components/schemas/ChainId"
        token:
          $ref: "#/components/schemas/Token"
        externalContentId:
          type: string
          description: type=purchase 时必填
        price:
          type: number
          minimum: 0.01
          maximum: 10000
          description: type=purchase 时必填
        amount:
          type: number
          minimum: 0.01
          maximum: 10000
          description: type=subscribe 和 type=tip 时必填
        externalTransactionId:
          type: string
          description: 所有类型（purchase、subscribe、tip）均必填。你的外部交易标识符。
...
paths:
...
  /v1/external/payment-intents:
    post:
      tags: [支付意图]
      summary: 创建支付意图
      description: |
        创建服务端支付意图。返回一个 intentId，可传递给弹窗 URL
        代替敏感参数（价格、收益方案等）。
        意图在 30 分钟后过期且仅可使用一次。
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentIntentRequest"
      responses:
        "201":
          description: 支付意图已创建
          content:
            application/json:
              schema:
                type: object
                properties:
                  intentId:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
        "400":
          description: 校验错误
        "401":
          description: API 密钥无效
...